brahma_hq_publisher/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-storage==2.18.2
pydantic==2.8.2
